refactor: installing new modules, flat routes
