+++ app/components/Alphabet/LexicalTreasuryContainer.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
 
 // === Imports ===
 import { AlphabetItem } from "~/types";
@@ -7,6 +6,8 @@ import { AlphabetLayout } from "./AlphabetLayout";
 import DisplayGroupMetadata from "./DisplayGroupMetadata";
 import { GroupingSelectForm } from "./GroupingSelectForm";
 import classificationData from "~/data/sanskrit-phoneme-groupings.json";
+import { useLocation } from "@remix-run/react";
+import { useNavigate } from "@remix-run/react";
 
 // === Type Definitions ===
 interface Props {
