+++ app/components/Explainer/ExplainerContainer.tsx
@@ -4,7 +4,7 @@ import { ArticleDisplay } from "~/components/Explainer/ArticleDisplay";
 import { ExplainerForm } from "~/components/Explainer/ExplainerForm";
 import { InstructionsBox } from "~/components/Explainer/InstructionsBox";
 import { SectionLoading } from "~/components/Shared/SectionLoading";
-import { explainConceptRequest } from "~/loader/explain-concept";
+import { explainConceptRequest } from "~/lib/loader/explain-concept";
 import { useConceptExplainer } from "~/hooks/concept-explainer";
 
 export default function ExplainerContainer() {
