+++ app/routes/api+/ai-storyteller-continue.tsx
