+++ app/routes/api+/ai-storyteller.tsx
