+++ app/routes/articles+/$id.tsx
@@ -2,15 +2,15 @@ import { type LoaderFunctionArgs } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { ArticleDetailContainer } from "~/components/Articles/ArticleDetailContainer";
 import { PageFrame } from "~/components/Layout/PageFrame";
-import { getArticleById } from "~/loader/articles";
+import { getArticleById } from "~/lib/loader/articles";
 
 
 export async function loader({ params }: LoaderFunctionArgs) {
-  const { slug } = params;
-  if (!slug) {
+  const { id } = params;
+  if (!id) {
     throw new Response("Not Found", { status: 404 });
   }
-  const article = await getArticleById(slug);
+  const article = await getArticleById(id);
   if (!article) {
     throw new Response("Not Found", { status: 404 });
   }
