+++ package.json
@@ -1,67 +1,106 @@
 {
-  "name": "sanskrit-garden",
+  "name": "my-remix-app",
   "private": true,
   "sideEffects": false,
   "type": "module",
   "scripts": {
     "build": "remix vite:build",
+    "preview": "remix vite:preview",
     "dev": "remix vite:dev",
     "lint": "eslint --ignore-path .gitignore --cache --cache-location ./node_modules/.cache/eslint .",
     "start": "remix-serve ./build/server/index.js",
     "typecheck": "tsc",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
+    "test:coverage": "jest --coverage",
+    "storybook": "STORYBOOK=true storybook dev -p 6006",
+    "storybook:build": "STORYBOOK=true storybook build",
+    "storybook:preview": "npx serve storybook-static"
   },
   "dependencies": {
     "@heroicons/react": "^2.2.0",
     "@remix-run/node": "^2.16.5",
     "@remix-run/react": "^2.16.5",
     "@remix-run/serve": "^2.16.5",
+    "@tailwindcss/typography": "^0.5.16",
+    "@tanstack/react-table": "^8.21.3",
+    "bootstrap": "^5.3.5",
     "chart.js": "^4.4.9",
     "chartjs-plugin-datalabels": "^2.2.0",
-    "clsx": "^2.1.1",
-    "isbot": "^4.1.0",
-    "marked": "^15.0.9",
-    "openai": "^4.96.0",
-    "react": "^18.2.0",
+    "d3": "^7.9.0",
+    "firebase": "^11.6.1",
+    "firebase-admin": "^13.3.0",
+    "gray-matter": "^4.0.3",
+    "heroicons": "^2.2.0",
+    "isbot": "^4.4.0",
+    "keen-slider": "^6.8.6",
+    "marked": "^15.0.8",
+    "openai": "^4.96.2",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.3.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
+    "react-hot-toast": "^2.5.2",
     "react-markdown": "^10.1.0",
+    "react-router-dom": "^7.6.0",
     "react-tooltip": "^5.28.1",
     "rehype-raw": "^7.0.0",
-    "remark": "^15.0.1",
-    "remark-breaks": "^4.0.0"
+    "remark-breaks": "^4.0.0",
+    "remark-gfm": "^4.0.1",
+    "vite": "^6.3.5",
+    "vite-plugin-ssr": "^0.4.142",
+    "zod": "^3.24.4"
   },
   "devDependencies": {
+    "@chromatic-com/storybook": "^3.2.6",
     "@remix-run/dev": "^2.16.5",
-    "@tailwindcss/typography": "^0.5.16",
+    "@storybook/addon-essentials": "^8.6.12",
+    "@storybook/addon-onboarding": "^8.6.12",
+    "@storybook/blocks": "^8.6.12",
+    "@storybook/experimental-addon-test": "^8.6.12",
+    "@storybook/react": "^8.6.12",
+    "@storybook/react-vite": "^8.6.12",
+    "@storybook/test": "^8.6.12",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
+    "@types/d3": "^7.4.3",
     "@types/jest": "^29.5.14",
-    "@types/mocha": "^10.0.10",
-    "@types/react": "^18.2.20",
+    "@types/react": "^18.3.20",
     "@types/react-dom": "^18.2.7",
-    "@types/testing-library__react": "^10.0.1",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
     "@typescript-eslint/parser": "^6.7.4",
+    "@vitest/browser": "^3.1.3",
+    "@vitest/coverage-v8": "^3.1.3",
     "autoprefixer": "^10.4.19",
     "eslint": "^8.38.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-import-resolver-typescript": "^3.10.1",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-storybook": "^0.12.0",
+    "fs-extra": "^11.3.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "playwright": "^1.52.0",
     "postcss": "^8.4.38",
+    "remix-flat-routes": "^0.8.5",
+    "remix-ssg": "^0.0.8",
+    "serve": "^14.2.4",
+    "storybook": "^8.6.12",
     "tailwindcss": "^3.4.4",
     "ts-jest": "^29.3.2",
-    "typescript": "^5.1.6",
-    "vite": "^6.0.0",
-    "vite-tsconfig-paths": "^4.2.1"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3",
+    "vite-tsconfig-paths": "^4.2.1",
+    "vitest": "^3.1.3"
   },
   "engines": {
     "node": ">=20.0.0"
-  }
+  },
+  "version": "1.0.0",
+  "description": "- 📖 [Remix docs](https://remix.run/docs)",
+  "main": "postcss.config.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
